+++ tests/testthat/test-bm_trim.R
@@ -1,7 +1,8 @@
+font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
+font <- read_hex(font_file)
+capital_r <- font[[str2ucp("R")]]
+
 test_that("bm_trim()", {
-    font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
-    font <- read_hex(font_file)
-    capital_r <- font[[str2ucp("R")]]
     capital_r_trimmed <- bm_trim(capital_r, c(1, 1, 3, 0))
     expect_equal(nrow(capital_r_trimmed), 12L)
     expect_equal(ncol(capital_r_trimmed), 7L)
@@ -32,4 +33,12 @@ test_that("bm_trim()", {
                   print(capital_r_resized2, labels = c("-", "#")))
     expect_equal(ncol(capital_r_resized2), 7L)
     expect_equal(nrow(capital_r_resized2), 18L)
+})
+
+test_that("bm_shift()", {
+    capital_r_shifted <- bm_shift(capital_r, bottom = 2L)
+    verify_output("txt/capital_r_shifted.txt",
+                  print(capital_r_shifted, labels = c("-", "#")))
+    expect_equal(nrow(capital_r_shifted), 16L)
+    expect_equal(ncol(capital_r_shifted), 8L)
 })